add: wake up assistant by keyboard hotkey
Default hotkey is `windows+shift+a`. Need to see if it doesn't
conflict with some other system wide default hotkey.

I've tested it for OSX

It's a little weird since the right command key and left shift key
doesn't work. So the hotkey has to be left command key + right shift
key + a